[Chore] 삭제된 데이터 getAll 로직 제거
