+++ src/main/java/com/example/real_chat/repository/RoomRepository.java
@@ -11,7 +11,6 @@ public interface RoomRepository {
 
     Optional<ChatRoom> findById(Long id);
     List<ChatRoom> findAll();
-    List<ChatRoom> findUnDeletedRooms();
 
     void delete(ChatRoom chatRoom);
 }
