+++ src/main/java/com/example/real_chat/repository/UserRepository.java
@@ -11,7 +11,6 @@ public interface UserRepository {
 
     Optional<User> findById(Long id);
     List<User> findAll();
-    List<User> findUnDeletedUsers();
 
     void delete(User user);
 }
