+++ src/main/java/com/example/real_chat/service/query/RoomQueryService.java
@@ -8,5 +8,4 @@ public interface RoomQueryService {
 
     ChatRoom getRoom(Long roomId);
     List<ChatRoom> getAllRooms();
-    List<ChatRoom> getUnDeletedRooms();
 }
